Simplify Makefiles embedded in autotest.at
This is a follow-up for the various patches to address problems with
tests 221 and 222 with various non-GNU make implementations.  We’re
not trying to exercise Make at all in these tests; it’s just a
convenient way to invoke the compiler found by AC_PROG_CC on a test
program.  The tests will be more reliable if we minimize the number of
Make features we are using.  So: no implicit rules at all, and no
intermediates.  Generate ‘testprog’ directly from ‘testprog.c’.

Also copy from fortran.at a more thorough set of substitution
variables for the compilation command, mainly for consistency,
and don’t use Makefile variables, again for consistency with
fortran.at.

(This is also, theoretically, faster since we’re only invoking the
compiler driver once, but it’s probably not enough of a difference to
measure.)