AC_PROG_*CPP: Try ‘cpp’ before ‘/lib/cpp’
All of the AC_PROG_*CPP macros fall back to ‘/lib/cpp’ if invoking the
relevant compiler with ‘-E’ doesn’t work.  This is an obsolete
location to find the C preprocessor.  Linux distributions tend to ship
a compatibility symlink, as do Solaris and AIX (as of the versions in the
GCC compile farm, anyway), but the BSDs no longer do.

The tests for this fallback are already expecting ‘cpp’ to be a
command findable via $PATH, so make the probe macros do the same
thing.  (We still try ‘/lib/cpp’ if ‘cpp’ _isn’t_ in $PATH.)

* lib/autoconf/c.m4
  (AC_PROG_CPP, AC_PROG_CXXCPP, AC_PROG_OBJCPP, AC_PROG_OBJCXXCPP):
  If $CC / $CXX / $OBJC / $OBJCXX -E (respectively) does not run the
  preprocessor, look for ‘cpp’ in $PATH before falling back to ‘/lib/cpp’.

* doc/autoconf.texi
  (AC_PROG_CPP, AC_PROG_CXXCPP, AC_PROG_OBJCPP, AC_PROG_OBJCXXCPP):
  Update to match.