Fetch more files from their official upstream.
‘make fetch’ was updating several auxiliary files from the copy
embedded in Gnulib instead of from their official home.  This patch
teaches fetch.pl to update as many as possible of them from their
official homes.  The remaining exception is GNU license files, which,
except for fdl.texi, do not appear to have a public version control
system.  Gnulib has a script for updating them, but it appears one
must be logged into some specific FSF-operated shell host for it to
work.

The files affected, and their official homes, are:

  build-aux/install-sh    automake
  build-aux/mdate-sh      automake
  build-aux/config.guess  config
  build-aux/config.sub    config
  doc/fdl.texi            gnustandards
  doc/gnu-oids.texi       gnustandards
  doc/make-stds.texi      gnustandards
  doc/standards.texi      gnustandards
  doc/texinfo.tex         texinfo

This change exposed a bug in fetch.sh: there is a subtle difference
between File::Spec->catpath and ->catfile and I should have been using
the latter.

* build-aux/fetch.pl: Can now fetch from arbitrary Savannah
  repositories, not just from gnulib and automake.  Fetch as
  many files as possible from their official home, instead of
  from gnulib.  Use File::Spec->catfile, not ->catpath.
  Strip trailing whitespace from all files.