Rewrite mktests.sh in Perl.
This is probably marginally faster since it doesn’t have to read the
.m4 files twice or muck around with temporary files and subprocesses,
but the actual point of this rewrite is that it will make it easier to
provide additional arguments to AT_CHECK_MACRO and AT_CHECK_AU_MACRO
on a per-macro basis; this capability will be added and used in the
next patch.

In *this* patch, the ac*.at files are not functionally changed at all.
(The comments come out slightly differently, though.)

* tests/mktests.sh: Delete and...
* tests/mktests.pl: ...rewrite in Perl.
* tests/local.mk: Update to match.